* added check for correct python version, prints proper error message when executed from older python version (#61)
* refactored tests to that test_versioncheck.py also works with python 2.7
