+++ .travis.yml
@@ -5,6 +5,8 @@ install: pip install -U tox coveralls
 matrix:
   fast_finish: true
   include:
+    - python: "2.7"
+      env: TOXENV=py27
     - python: "3.4.2"
       env: TOXENV=py34
     - python: "3.5"
