+++ test/unit_tests/test_miflora_poller.py
@@ -2,7 +2,7 @@
 from miflora.miflora_poller import MiFloraPoller, MI_LIGHT, MI_TEMPERATURE, MI_MOISTURE, MI_CONDUCTIVITY, MI_BATTERY
 
 import unittest
-from test.unit_tests import MockBackend
+from test.helper import MockBackend
 from test import HANDLE_WRITE_MODE_CHANGE, HANDLE_READ_SENSOR_DATA, INVALID_DATA
