+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py34, py35, py36, pylint, flake8, integration_tests
+envlist = py27,  py34, py35, py36, pylint, flake8, integration_tests
 skip_missing_interpreters = True
 deps = -rrequirements.txt
 
@@ -8,6 +8,11 @@ deps = -rrequirements.txt
 deps= -rrequirements.txt
 passenv = TRAVIS TRAVIS_*
 commands = pytest --cov=miflora test/unit_tests
+
+[testenv:py27]
+# only run unit tests as they do not need additional hardware
+deps= -rrequirements.txt
+commands = pytest test/unit_tests/test_versioncheck.py
 
 [testenv:integration_tests]
 #there tests are run separately as they rquire read hardware
